requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.2
langchain-ollama==0.1.0rc0
langchain-core==0.2.43 